+++ index.js
@@ -4,8 +4,15 @@ const cors = require('cors')
 const helmet = require('helmet')
 const server = express()
 const routes = require('./routes')
-const { notFoundHandler, logErrors, errorHandler } = require('./middlewares')
-const { PORT, ENVIRONMENT } = require('./config')
+const {
+  notFoundHandler,
+  logErrors,
+  boomErrorHandler,
+  ormErrorHandler,
+  errorHandler
+} = require('./middlewares')
+const { PORT, ENVIRONMENT } = require('./config/server')
+const db = require('./db/models')
 
 server.use(cors())
 server.use(helmet())
@@ -14,10 +21,20 @@ server.use(express.urlencoded({ extended: false }))
 server.use('/api/v1', routes)
 server.use(notFoundHandler)
 server.use(logErrors)
+server.use(boomErrorHandler)
+server.use(ormErrorHandler)
 server.use(errorHandler)
 
-if (ENVIRONMENT !== 'test') {
-  server.listen(PORT, () => console.log(`Server running in port ${PORT}`))
-} else {
-  module.exports = server
-}
+;(async () => {
+  try {
+    if (ENVIRONMENT === 'test') {
+      module.exports = server
+    } else {
+      await db.sequelize.sync().then(() => console.log('Sync DB'))
+      server.listen(PORT, () => console.log(`Server running in port ${PORT}`))
+    }
+  } catch (err) {
+    console.error(err)
+    process.exit(1)
+  }
+})()