+++ routes/brands.router.js
@@ -7,10 +7,11 @@ const {
   createBrandSchema,
   createModelByBrandIdSchema
 } = require('../schemas/brands.schema')
+const { brandsSequelizeRepository } = require('../repositories')
 const router = express.Router()
 
 router.get('/', asyncHandler(async (req, res) => {
-  const brands = await brandsServices.findBrands()
+  const brands = await brandsServices(brandsSequelizeRepository).findBrands()
   res.status(200).json(brands)
 }))
 
@@ -18,7 +19,7 @@ router.get('/:id/models',
   validatorHandler(findBrandIdSchema, 'params'),
   asyncHandler(async (req, res) => {
     const { id } = !!req.params && req.params
-    const brand = await brandsServices.findModelsByBrandId(id)
+    const brand = await brandsServices(brandsSequelizeRepository).findModelsByBrandId(id)
     res.status(200).json(brand)
   }))
 
@@ -26,7 +27,7 @@ router.post('/',
   validatorHandler(createBrandSchema, 'body'),
   asyncHandler(async (req, res) => {
     const { body } = !!req && req
-    const newBrand = await brandsServices.createBrand(body)
+    const newBrand = await brandsServices(brandsSequelizeRepository).createBrand(body)
     res.status(201).json(newBrand)
   }))
 
@@ -36,7 +37,7 @@ router.post('/:id/models',
   asyncHandler(async (req, res) => {
     const { params, body } = !!req && req
     const { id } = !!params && params
-    const newModel = await brandsServices.createModelByBrandId(id, body)
+    const newModel = await brandsServices(brandsSequelizeRepository).createModelByBrandId(id, body)
     res.status(201).json(newModel)
   }))