+++ routes/models.router.js
@@ -3,13 +3,14 @@ const asyncHandler = require('express-async-handler')
 const { modelsServices } = require('../services')
 const validatorHandler = require('../middlewares/validator.handler')
 const { findModelIdSchema, findModels, updateModelSchema } = require('../schemas/models.schema')
+const { modelsSequelizeRepository } = require('../repositories')
 const router = express.Router()
 
 router.get('/',
   validatorHandler(findModels, 'query'),
   asyncHandler(async (req, res) => {
     const { query } = !!req && req
-    const models = await modelsServices.findModels(query)
+    const models = await modelsServices(modelsSequelizeRepository).findModels(query)
     res.status(200).json(models)
   }))
 
@@ -20,7 +21,7 @@ router.put('/:id',
     const { body, params } = !!req && req
     const { id } = !!params && params
 
-    const updateModel = await modelsServices.updateModel(id, body)
+    const updateModel = await modelsServices(modelsSequelizeRepository).updateModel(id, body)
     res.status(200).json(updateModel)
   }))