+++ services/brands.service.js
@@ -1,67 +1,52 @@
 const boom = require('@hapi/boom')
-const { brands, models } = require('../lib/mock-db')
-const { searchModelsByBrandId, searchName, calculateId, searchIndex } = require('../helpers')
 
-async function findBrands () {
-  return brands || []
-}
-
-async function findModelsByBrandId (id) {
-  const brand = searchModelsByBrandId(id, models)
-
-  if (!brand) {
-    throw boom.notFound('brand not found')
+module.exports = brandsRepository => {
+  async function findBrands () {
+    return await brandsRepository.findBrands()
   }
 
-  return brand
-}
+  async function findModelsByBrandId (id) {
+    const brand = await brandsRepository.findBrandById(id)
 
-async function createBrand (data) {
-  const { name } = data
-  const nameExists = searchName(name, brands)
+    if (!brand) {
+      throw boom.notFound('brand not found')
+    }
 
-  if (nameExists) {
-    throw boom.conflict('name brand exist')
+    return await brandsRepository.findModelsByBrandId(id)
   }
 
-  const id = calculateId(brands)
-  const newBrand = { id, average_price: 0, ...data }
-  brands.push(newBrand)
+  async function createBrand (data) {
+    const { name } = !!data && data
+    const nameExists = await brandsRepository.findBrandByName(name)
 
-  return newBrand
-}
+    if (nameExists) {
+      throw boom.conflict('name brand exist')
+    }
 
-async function createModelByBrandId (brandId, data) {
-  const { average_price } = !!data && data
-  const brandIndex = searchIndex(brandId, brands)
-
-  if (brandIndex === -1) {
-    throw boom.notFound('brand not exist')
+    return await brandsRepository.createBrand(data)
   }
 
-  const { name } = data
-  const nameExists = searchName(name, models)
+  async function createModelByBrandId (brandId, data) {
+    const { name } = !!data && data
+    const brand = await brandsRepository.findBrandById(brandId)
 
-  if (nameExists) {
-    throw boom.conflict('name model exist')
+    if (!brand) {
+      throw boom.notFound('brand not found')
+    }
+
+    const nameModelExists = await brandsRepository.findModelByName(name)
+
+    if (nameModelExists) {
+      throw boom.conflict('name model exist')
+    }
+
+    return await brandsRepository.createModelByBrandId(brandId, data)
   }
 
-  // Create model
-  const id = calculateId(models)
-  const newModel = { id, brandId: +brandId, ...data }
-  models.push(newModel)
-
-  // Update average_price in brand
-  const brand = brands[brandIndex]
-  const newAveragePrice = (brand?.average_price || 0) + (average_price || 0)
-  brands[brandIndex] = { ...brand, average_price: newAveragePrice }
-
-  return newModel
-}
-
-module.exports = {
-  findBrands,
-  findModelsByBrandId,
-  createBrand,
-  createModelByBrandId
+  return {
+    findBrands,
+    findModelsByBrandId,
+    createBrand,
+    createModelByBrandId
+  }
 }