+++ CHANGELOG.md
@@ -2,6 +2,56 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.0.0](https://github.com/jousmo/nexu-backend-test/compare/v1.0.0...v2.0.0) (2022-07-29)
+
+
+### Features
+
+* add new script for test and init database ([409028f](https://github.com/jousmo/nexu-backend-test/commit/409028fe21defd1aac3f20cba8ecffe40890a8ad))
+* association brands and models sequelize ([12f96d9](https://github.com/jousmo/nexu-backend-test/commit/12f96d95f739f153a106e4249eeadc569d86c3b7))
+* config database connection for sequelize ([c045e81](https://github.com/jousmo/nexu-backend-test/commit/c045e818239a46ecce3f78e709b1b11e7aab46a6))
+* configuration instance sequelize ([b3858f1](https://github.com/jousmo/nexu-backend-test/commit/b3858f1d46267c5590d3467caa1c26802ff10ed6))
+* configure paths sequelize for cli ([020c9b8](https://github.com/jousmo/nexu-backend-test/commit/020c9b83432cc5ed36a347a524b2bf65974ca276))
+* create build models ([608fedc](https://github.com/jousmo/nexu-backend-test/commit/608fedc2e34db9e5e7133ee57f2c94e37beec2d0))
+* generate seeder for brands and models ([d853c7b](https://github.com/jousmo/nexu-backend-test/commit/d853c7bf1eebbc1aa2c3fc942d942e2c3f20d5fd))
+* implement all brand services with mock and sequelize repository ([4b1e0dd](https://github.com/jousmo/nexu-backend-test/commit/4b1e0ddd835b88b961f0d44a07452b940a0d7fda))
+* implement brands and models, model sequelize ([8ff753f](https://github.com/jousmo/nexu-backend-test/commit/8ff753f7898b19aa2b8a2aabd22183c6b9ac9a27))
+* implement docker compose for postgres local database ([fa02ef7](https://github.com/jousmo/nexu-backend-test/commit/fa02ef7d1122b5da9cd6601a78ff37a8b24dbcc9))
+* implement logic sequelize repository ([fec2979](https://github.com/jousmo/nexu-backend-test/commit/fec29798ffc70a2779c669ac14a5533f9ccc76de))
+* implement repositories for brands and models ([e81e679](https://github.com/jousmo/nexu-backend-test/commit/e81e679b76db93d4057fe9779dead39a6816c952))
+* include test config for sequelize ([d1e6c62](https://github.com/jousmo/nexu-backend-test/commit/d1e6c62e42bc12840faa5cc895265fd6bfded1ca))
+* install package sequelize ([789a4a3](https://github.com/jousmo/nexu-backend-test/commit/789a4a33a2ea017deb8c4f79caf7f4e06aa42dbc))
+* install packages cli's dotenv, sequelize and package dayjs and cross var ([2fd78bd](https://github.com/jousmo/nexu-backend-test/commit/2fd78bd8a1741dc26db2f986a2418d9a3281b53b))
+* separate logs error ([454e9db](https://github.com/jousmo/nexu-backend-test/commit/454e9dbb887c450b1f257e602c66a40d9107f63d))
+* update brand mock repository ([e56205a](https://github.com/jousmo/nexu-backend-test/commit/e56205a15fa1a107a83609dc67ea8a0318d7560c))
+* update config vars separate server and sequelize ([f4dfaa1](https://github.com/jousmo/nexu-backend-test/commit/f4dfaa178415917a707c21189ba7883795861076))
+* update documentation, init postgres database and sequelize create db for test ([e9b479b](https://github.com/jousmo/nexu-backend-test/commit/e9b479bb1b02131df0ddeeb4ea7bbbd859a2c2cd))
+* update env vars ([b07c60d](https://github.com/jousmo/nexu-backend-test/commit/b07c60d9cbc288631f172c3ce22524fd360be516))
+* update env vars and connection sequelize ([d9297ed](https://github.com/jousmo/nexu-backend-test/commit/d9297ed8df8bc41bd844500eab44080c14ebeb86))
+* update env vars for config database ([646b508](https://github.com/jousmo/nexu-backend-test/commit/646b508507962903f1ce6fcee4ee46b0e4393e38))
+* update error handler support database errors ([6d0c4ac](https://github.com/jousmo/nexu-backend-test/commit/6d0c4ac96950a7e8b48453d4675b4c8325fc6f83))
+* update models mock repository ([d88396a](https://github.com/jousmo/nexu-backend-test/commit/d88396aef24eb00b231785ac5c3f6fd9463caf77))
+* update path config and sync database before server listening ([3014e95](https://github.com/jousmo/nexu-backend-test/commit/3014e95811ec90b3bda372432db93b53f968b235))
+* update readme ([b882161](https://github.com/jousmo/nexu-backend-test/commit/b882161596ad718eada19f58dfa273facfa8014d))
+* update services implementation with dependencies injections ([2709b20](https://github.com/jousmo/nexu-backend-test/commit/2709b2008dac4bbbd5025a3e1355512d8de5aecb))
+* update unit test ([436d692](https://github.com/jousmo/nexu-backend-test/commit/436d69241b7b55a7f9fc14d83f38ec52a2d6c0ac))
+* use dependencies injections in the services to use brands and model repositories ([6f7762b](https://github.com/jousmo/nexu-backend-test/commit/6f7762b557d832dccd5a68af89bb23ba843f918c))
+* use model service with mock o sequelize repository ([59ed6de](https://github.com/jousmo/nexu-backend-test/commit/59ed6dea92beb19ed5ae0f3f2e548285d62e6187))
+
+
+### Bug Fixes
+
+* change logic mock repository ([13caf7d](https://github.com/jousmo/nexu-backend-test/commit/13caf7d547ce23b38885dcef0b2215eb3b0d94fe))
+* create brand only data ([d030666](https://github.com/jousmo/nexu-backend-test/commit/d030666d7669cd1c6650e98b9f7ba30a77d276af))
+* export repositories ([34fa617](https://github.com/jousmo/nexu-backend-test/commit/34fa617d0399c60cc6802939bd1cbeaa119db477))
+* implement middleware for all errors ([febf396](https://github.com/jousmo/nexu-backend-test/commit/febf396fd2120b4704b747766b384d4f68517245))
+* implement sequelize brands repository logic ([1da3b22](https://github.com/jousmo/nexu-backend-test/commit/1da3b2272f4900b3b7be2ab740fb098ffa9cf9cc))
+* method search models by brand mock helper ([47288b1](https://github.com/jousmo/nexu-backend-test/commit/47288b1aac32e6b6fbeed2506149737156b8fa09))
+* remove id for seeders ([98f6d6c](https://github.com/jousmo/nexu-backend-test/commit/98f6d6c41b28b1668a8ef1e61cc7884a5a5e98d9))
+* support config for test and development sequelize ([f090d06](https://github.com/jousmo/nexu-backend-test/commit/f090d06e08438c5d6d4c8922c183484d362c3642))
+* update server export only test ([087da95](https://github.com/jousmo/nexu-backend-test/commit/087da951f5585b7ac11e90e6e6d68b299350f14c))
+* use repositories in services with data injection ([e0947c3](https://github.com/jousmo/nexu-backend-test/commit/e0947c3e4ed6c6fe5948f1986c4c3f8377835b1d))
+
 ## [1.0.0](https://github.com/jousmo/nexu-backend-test/compare/v0.2.0...v1.0.0) (2022-07-27)